refactor(naming) rename Modals to Forms
