+++ src/components/Pages/Members/Members.jsx
@@ -3,9 +3,9 @@ import { PageTitle } from '../../PageTitle/PageTitle';
 import { TABLE_TITLES, TITLES_PAGES, BUTTONS_NAMES, LINKPATH_KEYS, MODALTITLE_KEYS } from '../../../shared/constants';
 import { TableMembers } from '../../Table/TableMembers';
 import { Modal } from '../../Common/Modal/Modal';
-import { CreateMemberModal } from '../../Common/Modal/CreateMemberModal/CreateMemberModal';
 import { getAllUsers } from '../../../services/users-services ';
 import { ButtonsAdminMemberPage } from '../../Buttons/ButtonsAdmin/ButtonsAdmin';
+import { CreateMemberForm } from '../../Forms/CreateMemberForm/CreateMemberForm';
 
 export class Members extends React.Component {
   constructor(props) {
@@ -47,7 +47,7 @@ export class Members extends React.Component {
             isModalOpen={isModalOpen}
             toggleModalHandler={this.toggleModalHandler}
           >
-            <CreateMemberModal toggleModalHandler={this.toggleModalHandler} setUsersHandler={this.setUsersHandler} />
+            <CreateMemberForm toggleModalHandler={this.toggleModalHandler} setUsersHandler={this.setUsersHandler} />
           </Modal>
         )}
       </>
